Move client to launchpad
