+++ README.md
@@ -2,9 +2,8 @@
 
 |Package|README|
 |:----:|:---:|
-| [![devtools-client-adapters-version]][devtools-client-adapters-pkg] | [devtools-client-adapters] |
+
 | [![devtools-launchpad-version]][devtools-launchpad-pkg] | [devtools-launchpad] |
-| [![devtools-network-request-version]][devtools-network-request-pkg] | [devtools-network-request] |
 | [![devtools-config-version]][devtools-config-pkg] | [devtools-config] |
 | [![devtools-modules-version]][devtools-modules-pkg] | [devtools-modules] |
 | [![devtools-sham-modules-version]][devtools-sham-modules-pkg] | [devtools-sham-modules] |
@@ -19,17 +18,9 @@
   - [Update a Package](./docs/versioning.md#update-a-package)
   - [Updating a Launchpad Dependency](./docs/versioning.md#updating-a-launchpad-dependency)
 
-[devtools-client-adapters-version]:https://img.shields.io/npm/v/devtools-client-adapters.svg
-[devtools-client-adapters-pkg]:https://npmjs.org/package/devtools-client-adapters
-[devtools-client-adapters]:./packages/devtools-client-adapters/#readme
-
 [devtools-launchpad-version]:https://img.shields.io/npm/v/devtools-launchpad.svg
 [devtools-launchpad-pkg]:https://npmjs.org/package/devtools-launchpad
 [devtools-launchpad]:./packages/devtools-launchpad/#readme
-
-[devtools-network-request-version]:https://img.shields.io/npm/v/devtools-network-request.svg
-[devtools-network-request-pkg]:https://npmjs.org/package/devtools-network-request
-[devtools-network-request]:./packages/devtools-network-request/#readme
 
 [devtools-config-version]:https://img.shields.io/npm/v/devtools-config.svg
 [devtools-config-pkg]:https://npmjs.org/package/devtools-config
