+++ docs/local-development.md
@@ -12,21 +12,21 @@ There are a couple strategies for this: link, local install, and package install
 #### Linking
 
 Linking is the most common approach.
-Here, yarn creates a symlink from the tools's `node_modules/devtools-client-adapters` directory to the
-devtools-core `devtools-client-adapters` package directory.
+Here, yarn creates a symlink from the tools's `node_modules/devtools-modules` directory to the
+devtools-core `devtools-modules` package directory.
 
 * Once the tool is linked then any change in the code will be reflected in the tool.
 * Making changes does not require re-starting the dev-server.
 * Ignores any pre-publish steps that might be important.
-* Could create a bundling issue because `devtools-client-adapters` has its own set of node_modules, which are not shared with the debugger or other tools.
+* Could create a bundling issue because `devtools-modules` has its own set of node_modules, which are not shared with the debugger or other tools.
 
-Steps for linking `devtools-client-adapters` and `debugger.html`:
+Steps for linking `devtools-modules` and `debugger.html`:
 
 ```bash
-cd devtools-core/packages/devtools-client-adapters
+cd devtools-core/packages/devtools-modules
 yarn link
 cd ../../../debugger.html
-yarn link devtools-client-adapters
+yarn link devtools-modules
 ```
 
 #### Local install
@@ -38,15 +38,15 @@ There are two options:
 #####  1. adding a local package
 ```bash
 cd debugger.html
-yarn add file:../devtools-core/packages/devtools-client-adapters
+yarn add file:../devtools-core/packages/devtools-modules
 ```
 
 ##### 2. adding a local tarball
 ```bash
-cd devtools-core/packages/devtools-client-adapters
+cd devtools-core/packages/devtools-modules
 yarn pack
 cd ../../../debugger.html
-yarn add ../devtools-core/packages/devtools-client-adapters/devtools-client-adapters-v0.0.5.tgz
+yarn add ../devtools-core/packages/devtools-modules/devtools-modules-v0.0.5.tgz
 ```
 
 The two options are very similar, but have slightly different semantics.
