--- packages/devtools-client-adapters/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "devtools-client-adapters",
-  "version": "0.0.24",
-  "description": "Devtools Client Adapters",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "dependencies": {
-    "chrome-remote-interface": "0.17.0",
-    "devtools-config": "^0.0.12",
-    "devtools-modules": "^0.0.18",
-    "devtools-network-request": "^0.0.14",
-    "devtools-sham-modules": "^0.0.18"
-  },
-  "files": [
-    "index.js",
-    "src"
-  ],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-  }
-}
