+++ packages/devtools-launchpad/package.json
@@ -49,12 +49,11 @@
     "body-parser": "^1.15.2",
     "check-node-version": "^1.1.2",
     "classnames": "^2.2.5",
+    "chrome-remote-interface": "0.17.0",
     "co": "=4.6.0",
     "css-loader": "^0.26.1",
-    "devtools-client-adapters": "^0.0.24",
     "devtools-config": "^0.0.12",
     "devtools-modules": "^0.0.18",
-    "devtools-network-request": "^0.0.14",
     "devtools-sham-modules": "^0.0.18",
     "eslint": "^3.12.0",
     "eslint-plugin-babel": "^3.3.0",
