+++ packages/devtools-modules/client/shared/privilegedNetworkRequest.js
