+++ packages/devtools-modules/client/shared/shim/networkRequest.js
@@ -3,16 +3,15 @@
 // configured.
 function networkRequest(url, opts) {
   return Promise.race([
-    fetch(`/get?url=${url}`)
-      .then(res => {
-        if (res.status >= 200 && res.status < 300) {
-          return res.text().then(text => ({ content: text }));
-        }
-        return Promise.reject(new Error(`failed to request ${url}`));
-      }),
+    fetch(`/get?url=${url}`).then(res => {
+      if (res.status >= 200 && res.status < 300) {
+        return res.text().then(text => ({ content: text }));
+      }
+      return Promise.reject(new Error(`failed to request ${url}`));
+    }),
     new Promise((resolve, reject) => {
       setTimeout(() => reject(new Error("Connect timeout error")), 6000);
-    })
+    }),
   ]);
 }
