+++ packages/devtools-modules/index.js
@@ -1,3 +1,4 @@
+const networkRequest = require("./client/shared/shim/networkRequest");
 const Services = require("./client/shared/shim/Services");
 const SplitBox = require("./client/shared/components/splitter/SplitBox");
 // const SplitBoxCSS = require("./client/shared/components/splitter/SplitBox.css")
@@ -5,6 +6,7 @@ const sprintf = require("./shared/sprintf").sprintf;
 const workerUtils = require("./shared/worker-utils");
 
 module.exports = {
+  networkRequest,
   Services,
   SplitBox,
   // SplitBoxCSS,
