--- packages/devtools-network-request/README.md
@@ -1,3 +1,0 @@
-## Devtools Network Request
-
-[![Npm version](https://img.shields.io/npm/v/devtools-network-request.svg)](https://npmjs.org/package/devtools-network-request)
