--- packages/devtools-network-request/index.js
@@ -1,3 +1,0 @@
-const networkRequest = require("./networkRequest");
-
-module.exports = networkRequest;
