--- packages/devtools-network-request/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "devtools-network-request",
-  "version": "0.0.14",
-  "description": "Devtools Library for making cross origin requests",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "devDependencies": {
-    "devtools-config": "^0.0.12"
-  },
-  "author": "",
-  "license": "ISC"
-}
