+++ packages/devtools-source-map/package.json
@@ -15,8 +15,6 @@
   "dependencies": {
     "babel-plugin-transform-async-to-generator": "^6.22.0",
     "babel-plugin-transform-flow-strip-types": "^6.22.0",
-    "devtools-client-adapters": "^0.0.24",
-    "devtools-network-request": "^0.0.14",
     "devtools-config": "^0.0.12",
     "devtools-modules": "^0.0.18",
     "jest": "^19.0.2",
